# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API - HackSOAT Created/Validate User",
    "description": "Documentation of HackSOAT"    
  },
 

  "paths": {   
    "/create-user": {
      "post": {  
        "tags":["USER"],   
        "description": "",
        "requestBody": {
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "username":{
                    "type":"string"                    
                  },
                  "email":{
                    "type":"string"
                  },
                  "password":{
                    "type":"string"
                  }
                  
                }
              }
            }
          }
        },
        "responses": {
          "201":{
            "description":"Successfull"
          },
          "401": {
            "description": "Unautorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    
    "/auth":{
      "post": {
        "tags":["USER"], 

        "requestBody": {
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "username":{
                    "type":"string"                    
                  },                  
                  "password":{
                    "type":"string"
                  }
                  
                }
              }
            }
          }
        },
        "200":{
          "description":"Successfull"
        },
        "responses": {
          "401": {
            "description": "Unautorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
          
        }
      }
    },

    "/validate-create-user":{
      "post": {
        "tags":["USER"], 

        "requestBody": {
          "required":true,
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "username":{
                    "type":"string"                    
                  },                  
                  "confirmationCode":{
                    "type":"string"
                  }
                  
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unautorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
          
        }
      }
    }       
  }
}